Added the necessary validations

+++ lib/tasks/Contest_Reset.rake
@@ -1,10 +1,14 @@
 task({:auction_reset => :environment}) do
-  Winner.all.each do |a_winner|
-    a_winner.destroy
+  if Winner.all != nil 
+    Winner.all.each do |a_winner|
+      a_winner.destroy
+    end
   end
 
-  Bid.all.each do |a_bid|
-    a_bid.destroy
+  if Bid.all != nil 
+    Bid.all.each do |a_bid|
+      a_bid.destroy
+    end
   end
 
   Seeker.all.each do |a_seeker|
